DO NOT MERGE: Temporary paths for local developent
Hello, if you seeing this commit message and you are not trying to run
run a local development version of the GOV.UK Accounts save-a-page
feature... Then something is probably wrong.

This should not be merged into master.

The GOV.UK Account Save a page feature relies on sending the user to
frontend, which then calls the accounts api.

At the end of the transaction, irrespective of if it was sucessful or
not, then we send the user back to the piece of guidance they started
on.

Currently all guidance we are placing the button on is rendered in
`government-frontend`. However in future this may not be the case.

On integration, staging or production, both `frontend` and
`government-frontend` share a domain (www.gov.uk) and so can be
redirected purely by relative path.

However for the local docker setup we use subdomains
(`frontend.dev.gov.uk` and `government-frontend.dev.gov.uk`).

Without the changes in this commit, the redirects will try to throw a
user back to a path like
`frontend.dev.gov.uk/guidance/piece-of-guidance` which will fail because
the actual app we need to target is `government-frontend`.

However, the changes in this commit are too fragile to merge. If we
added the button to other apps, then `frontend` would end up trying to
redirect a development user to the wrong app.

So this branch, and this commit is purely a convenience for testing
purposes locally. If you find this commit in `master` or `main` in
future, or you did not intend for the behavious above, plese Drop this
commit or revert it.